pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.69.0",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "werkzeug>=3.1.3",
    "flask-wtf>=1.2.2",
    "humanize>=4.12.2",
    "sqlalchemy>=2.0.40",
    "trafilatura>=2.0.0",
    "wtforms>=3.2.1",
    "beautifulsoup4>=4.13.3",
    "requests>=2.32.3",
    "youtube-transcript-api>=1.0.3",
    "google-generativeai>=0.8.4",
    "pymupdf>=1.25.5",
    "pillow>=11.1.0",
    "playwright>=1.51.0",
    "cairosvg>=2.7.1",
]
